Make `get_post` useable without login
Before, the function would fail, because it requires `g.user` to be
something that can be accessed like a dictionary. If no user is logged
in, `g.user` would be `None`. This lead to errors when grabbing a post.

Added a check if `g.user is not None` before checking the equality of
post author and the current user. This allows to use the `get_post`
even when no user is logged in.