Fix not retrieving posts without tags
When the SQL was changed to retrieve the tags with the posts a bug was
introduced. The issue was how the tables where joined. The plain `JOIN`
would eliminate rows that did not have any values in the post-tag table.
This means posts without tags could not be found.

To fix this, the `LEFT JOIN` was used. This type of join will always
contain all rows from the "left" table. Rows where the join condition is
not met contain NULL values in the "right" tables columns.

This required an additional test to prevent the NULL value being
inserted when displaying existing tags in the tag edit input.