Whitespace and empty string handling for tags
The `get_or_create...` functions for tags did not have proper handling
for whitespace or empty strings so far. The tests and handling has been
added.

Before, it seems like that tags with empty string and white space had
been created in the database the first time the `get_or_create_tag`
function was called. This might also explain issues with not consistent
ids.