Updating tag association in form allows removal
It was not possible before to remove tags in the form and have the
association actually removed. The association would still exist in the
DB and thus be retrieved with the post.

This has been fixed with new function for updating the association of
tags with a post. This also reduced the dependency of the blog to one
function of the tags module.