Query definition and display changed
Before, there was no indication on the result page that it is filtered
for any value. Now the search bar is pre-filled with the term and the
headline shows the term too.

Also, the URL was changed to use the typical URL query string parameter
instead of a static looking URL.