Index limited to 5 posts per page
So far, only the reaction to the page URL query string parameter has
been set up. There are no controls on the index page to switch between
the pages yet.

Also, the number of posts per page is set in the index view function.
This should be a config setting.