Merged delete image option into post update form
Before, the delete image option was realized with its own URL and a
separate form on the post update page. This was not ideal, because
changes to the post properties like title and body would be lost when
the delete image form was submitted.

This is changed now, so that the delete image option is part of the
normal post update. The special URL for this option is not needed
anymore and was removed.