Sanitizing user input with markupsafe
Before, it was possible for users to submit html code in the blog posts.
This is a security issue! The only reason it did not become an immediate
problem was that the user content was added via Jinja to the pages.
Jinja does automatically escape the replacements in the templates, which
makes the html harmless (because the tag brackets are replaced with
`&lt;` and `&gt;`).

This sanitization/escaping through Jinja presented an issue when trying
to allow Markdown being displayed as HTML though. Because I would enter
the Markdown-converted-to-HTML per Jinja into the template the desired
HTML formatting would also be sanitized.

To stop Jinja from sanitizing the output if to use the `|safe` filter.
This would disable all of the sanitization though.

Thus, I needed a way to sanitize possible HTML in the user input before
transforming the Markdown to the desired HTML. This is where
`markupsafe.espace` comes in. With this function I am now sanitizing the
user input before it is even saved to the DB. This means it makes any
HTML injection useless as soon as the `create_or_update` view receives
it via HTTP POST.

This allows me to use the `|safe` filter in the Jinja template, because
the user input was sanitized before it was even written to the DB. This
in turn makes it possible to savely display the HTML generated from the
Markdown in the user input!